requirement.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.15.1
flask==3.0.3
flask-sqlalchemy==3.1.1
flatbuffers==24.3.25
fsspec==2024.6.0
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h5py==3.11.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
keras==2.13.1
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.30
sympy==1.12.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
torch==2.3.1
torchvision==0.18.1
triton==2.3.1
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
